+++ common/DSUtilLite/DShowUtil.cpp
@@ -491,15 +491,15 @@ void audioFormatTypeHandler(const BYTE *format, const GUID *formattype, DWORD *p
     *pnBytesPerSec = nBytesPerSec;
 }
 
-void getExtraData(const AM_MEDIA_TYPE &mt, BYTE *extra, unsigned int *extralen)
+void getExtraData(const AM_MEDIA_TYPE &mt, BYTE *extra, size_t *extralen)
 {
   return getExtraData(mt.pbFormat, &mt.formattype, mt.cbFormat, extra, extralen);
 }
 
-void getExtraData(const BYTE *format, const GUID *formattype, const size_t formatlen, BYTE *extra, unsigned int *extralen)
+void getExtraData(const BYTE *format, const GUID *formattype, const size_t formatlen, BYTE *extra, size_t *extralen)
 {
   const BYTE *extraposition = NULL;
-  unsigned extralength = 0;
+  size_t extralength = 0;
   if (*formattype == FORMAT_WaveFormatEx) {
     WAVEFORMATEX *wfex = (WAVEFORMATEX *)format;
     extraposition = format + sizeof(WAVEFORMATEX);
